Add user extraction middleware

+++ part4/bloglist/controllers/blogs.js
@@ -1,8 +1,8 @@
-const jwt = require('jsonwebtoken');
-
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
 const User = require('../models/user');
+
+const middleware = require('../middleware/middleware');
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', {
@@ -11,13 +11,8 @@ blogsRouter.get('/', async (request, response) => {
   response.json(blogs);
 });
 
-blogsRouter.post('/', async (request, response) => {
-  const decodedToken = jwt.verify(request.token, process.env.SECRET);
-  if (!decodedToken.id) {
-    response.status(401).json({ error: 'token invalid' });
-    return;
-  }
-  const user = await User.findById(decodedToken.id);
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const user = await User.findById(request.user);
   const blog = new Blog({ ...request.body, user: user._id });
   const savedBlog = await blog.save();
   user.blogs = user.blogs.concat(savedBlog._id);
@@ -25,7 +20,12 @@ blogsRouter.post('/', async (request, response) => {
   response.status(201).json(savedBlog);
 });
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id);
+  if (blog.user.toString() !== request.user.toString()) {
+    response.status(401).json({ error: 'Not authorized' });
+    return;
+  }
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
